Support Github Enterprise, upgrade Github client package
@octokit/rest is the new (since January) NPM package for accessing the
Github API. There are some differences that are taken care of here.

One of them is setting the mime type on asset upload.

Since octokit initialization happens on init (instead of module
execution), I've hacked up a means to use Sinon for mocking it.

The primary result of this change is to allow for publishing to Github
Enterprise instances.